+++ src/modules/bpmn-io/bpmn-io.html
@@ -1,19 +1,100 @@
 <template>
   <require from="../property-panel/property-panel"></require>
   <require from="@process-engine/bpmn-js-custom-bundle/bpmn-modeler-custom.css"></require>
+  <require from="../bpmn-xml-view/bpmn-xml-view"></require>
   <require from="./bpmn-io.css"></require>
-  <div class="modeler">
-    <div class="diagramm">
-      <div ref="canvasModel" class="canvasModel"></div>
+  <div class="window-layout">
+    <!--
+      DIAGRAM TOOLS
+    -->
+    <div class="window-layout__tools-left">
+      <div class="tool-list tool-list--vertical">
+        <div class="tool">
+          <i class="fa fa-mouse-pointer"></i>
+        </div>
+        <div class="tool btn-group">
+          <button class="menu-bar__menu-left--tool dropdown-toggle" title="Color Element" data-toggle="dropdown" click.delegate="updateCustomColors()">
+            <i class="fa fa-eyedropper"></i>
+          </button>
+          <ul class="dropdown-menu pull-left">
+            <li><a click.delegate="setColorRed()"><i class="glyphicon glyphicon-stop red"></i> Red</a></li>
+            <li><a click.delegate="setColorBlue()"><i class="glyphicon glyphicon-stop blue"></i> Blue</a></li>
+            <li><a click.delegate="setColorGreen()"><i class="glyphicon glyphicon-stop green"></i> Green</a></li>
+            <li><a click.delegate="setColorPurple()"><i class="glyphicon glyphicon-stop purple"></i> Purple</a></li>
+            <li><a click.delegate="setColorOrange()"><i class="glyphicon glyphicon-stop orange"></i> Orange</a></li>
+            <li>
+              <a show.bind="colorPickerLoaded" click.delegate="setColorPicked()">
+                <input ref="colorPickerFill"></input>
+                <input ref="colorPickerBorder"></input>
+              </a>
+            </li>
+            <li><a click.delegate="removeColor()"><i class="glyphicon glyphicon-stop none"></i> None</a></li>
+          </ul>
+        </div>
+        <div class="tool">
+          <button class="menu-bar__menu-left--tool dropdown-toggle" title="Distribute Elements Vertically" click.delegate="distributeElementsVertical()">
+            <i class="glyphicon glyphicon-object-align-vertical"></i>
+          </button>
+        </div>
+        <div class="tool">
+          <button class="menu-bar__menu-left--tool dropdown-toggle" title="Distribute Elements Vertically" click.delegate="distributeElementsHorizontal()">
+            <i class="glyphicon glyphicon-object-align-horizontal"></i>
+          </button>
+        </div>
+      </div>
+      <div class="tool-list tool-list--vertical tool-list--bottom">
+        <div class="tool">
+          <button class="menu-bar__menu-left--tool">
+            <a href="#/configuration" title="To Settings">
+              <i class="fa fa-cog"></i>
+            </a>
+          </button>
+        </div>
+      </div>
     </div>
-    <div class="resize_div" ref="resizeButton">
-      <button show.bind="!toggled" class="resize_button" title="Resize" click.delegate="resize()">Resize</button>
+
+    <!--
+      MAIN CANVAS
+    -->
+    <div class="window-layout__main js-main">
+      <div class="modeler" show.bind="!showXMLView">
+        <div class="diagramm">
+          <div ref="canvasModel" class="canvasModel"></div>
+        </div>
+        <!--<div class="resize_div" ref="resizeButton">
+          <button show.bind="!toggled" class="resize_button" title="Resize" click.delegate="resize()">Resize</button>
+        </div>
+        <div class="toggle_div" ref="toggleButton">
+          <button class="toggle_panel" title="Toggle PropertyPanel" click.delegate="togglePanel()">${ !toggled ? 'Hide' : 'Show'}</button>
+        </div>-->
+        <div class="properties" ref="panel">
+          <property-panel id="js-properties-panel" modeler.bind="modeler" xml.bind="xml"></property-panel>
+        </div>
+      </div>
+      <bpmn-xml-view if.bind="initialLoadingFinished" show.bind="showXMLView" xml.bind="xml"></bpmn-xml-view>
     </div>
-    <div class="toggle_div" ref="toggleButton">
-      <button class="toggle_panel" title="Toggle PropertyPanel" click.delegate="togglePanel()">${ !toggled ? 'Hide' : 'Show'}</button>
+
+    <div class="window-layout__tools-right">
+      <div class="tool-list tool-list--vertical">
+      </div>
     </div>
-    <div class="properties" ref="panel">
-      <property-panel id="js-properties-panel" modeler.bind="modeler" xml.bind="xml"></property-panel>
+    <!--
+    STATUS BAR
+    -->
+    <div class="status-bar status-bar--system-macos">
+      <div class="status-bar__left-bar">
+        <a class="status-bar__element status-bar__element--config-link" href="#/configuration">
+          <span>Connected to: ${env.bpmnStudioClient.baseRoute}</span>
+        </a>
+      </div>
+      <div class="status-bar__center-bar"></div>
+      <div class="status-bar__right-bar">
+        <div class="status-bar__element">
+          <button class="status-bar__show-xml-button" click.delegate="toggleXMLView()">
+            <i class="fa fa-file-code-o"></i> Show XML
+          </button>
+        </div>
+      </div>
     </div>
   </div>
 </template>
