+++ src/modules/property-panel/property-panel.scss
@@ -1,5 +1,7 @@
 .property-panel {
   width: 100%;
+  height: 100%;
+  overflow-y: scroll;
 }
 
 .indextab-dropdown {
